Changed implementation by replacing interface LibertyProjectMetadata with abstract class AbstractProjectMetadata
